+++ ddtrace/contrib/aiohttp_jinja2/__init__.py
@@ -20,8 +20,8 @@ required_modules = ["aiohttp_jinja2"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from ..internal.aiohttp_jinja2.patch import get_version
-        from ..internal.aiohttp_jinja2.patch import patch
-        from ..internal.aiohttp_jinja2.patch import unpatch
+        from ...internal.aiohttp_jinja2.patch import get_version
+        from ...internal.aiohttp_jinja2.patch import patch
+        from ...internal.aiohttp_jinja2.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
