+++ ddtrace/contrib/aiomysql/__init__.py
@@ -43,8 +43,8 @@ required_modules = ["aiomysql"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from ..internal.aiomysql.patch import get_version
-        from ..internal.aiomysql.patch import patch
-        from ..internal.aiomysql.patch import unpatch
+        from ...internal.aiomysql.patch import get_version
+        from ...internal.aiomysql.patch import patch
+        from ...internal.aiomysql.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
